+++ makefile
@@ -1,8 +1,8 @@
 
 all: main
 
-main: src/main.c
-	gcc src/main.c -o agogo
+main: src/main.c 
+	gcc src/main.c src/project.c src/clock.c  src/task.c -o agogo
 
 debug: src/main.c
 	gcc -g src/main.c -o agogo
