+++ src/agogo.h
@@ -1,0 +1,8 @@
+#ifndef AGOGO_H
+#define AGOGO_H
+
+#define AGOGO_VERSION "0.1"
+#define AGOGO_DIR "/home/mcromer/programming/c/agogo/tmp/agogo"
+#define AGOGO_PROJECTS_DIR AGOGO_DIR "/projects"
+
+#endif // AGOGO_H
