+++ src/clock.h
@@ -1,0 +1,8 @@
+#ifndef AGOGO_CLOCK_H
+#define AGOGO_CLOCK_H
+
+int agogo_clockon(int argc, char *argv[]);
+int agogo_clockoff();
+int is_clocked_on();
+
+#endif // AGOGO_CLOCK_H
