+++ src/project.h
@@ -1,0 +1,7 @@
+#ifndef PROJECT_H
+#define PROJECT_H
+
+int agogo_project(int argc, char *argv[]);
+char *get_current_project();
+
+#endif // PROJECT_H
