+++ src/task.h
@@ -1,0 +1,7 @@
+#ifndef AGOGO_TASK_H
+#define AGOGO_TASK_H
+
+int agogo_task(int argc, char *argv[]);
+void list_tasks();
+
+#endif // AGOGO_TASK_H
